+++ release-notes.md
@@ -1,4 +1,4 @@
 #Release 0.1.1
 ### major features & improvements
-* 
+* Klass to find classes in some packages
 ### buf fixes